Move email into namespace
