+++ src/FunderMaps.AspNetCore/HealthChecks/EmailHealthCheck.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using FunderMaps.Core.Interfaces;
+using FunderMaps.Core.Email;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Logging;
